+++ src/services/user.service.ts
@@ -11,16 +11,20 @@ import { getUserByIdSchema } from '../validation/user/getUserById.schema';
 import { updateUserSchema } from '../validation/user/updateUser.schema';
 import { validate } from '../validation/validate';
 
-export async function addUser(data: AddUserRequest): Promise<ServiceResult<AddedUserResponse>> {
+export async function addUserAsync(data: AddUserRequest): Promise<ServiceResult<AddedUserResponse>> {
    
   const validationResult = await validate(addUserSchema, data); 
   if (!validationResult.success) { 
-    return validationResult;
+    return { success: false, code: 400, error: validationResult.error }
   }  
   const validData = validationResult.data;
   
   try {   
    
+     if ((await usernameExistsAsync(validData.username))) {
+          return {success: false, code: 400, error: ['Username exists already']};
+        }   
+
     validData.password = await createHashedPassword(validData.password);
  
     const user = new UserModel(validData);
@@ -38,17 +42,22 @@ export async function addUser(data: AddUserRequest): Promise<ServiceResult<Added
   }
 } 
 
-export async function updateUser(data: UpdateUserRequest): Promise<ServiceResult<UpdatedUserResponse>> {
+export async function updateUserAsync(data: UpdateUserRequest): Promise<ServiceResult<UpdatedUserResponse>> {
    
   const validationResult = await validate(updateUserSchema, data);  
   if (!validationResult.success) {
-    return validationResult;
+    return { success: false, code: 400, error: validationResult.error }
   }   
 
   try {
 
     const { id: validId, surname: validSurname, firstName: validFirstName } = validationResult.data;
-
+ 
+    const existingUser = await UserModel.findById(validId);  
+    if (!existingUser) {
+      return {success: false, code: 404, error: ['User not found']};
+    } 
+ 
     const updatedUser = await UserModel.findByIdAndUpdate(
       validId,                          
       { $set: { surname: validSurname, firstName: validFirstName } },
@@ -56,7 +65,7 @@ export async function updateUser(data: UpdateUserRequest): Promise<ServiceResult
     );
 
     if (!updatedUser) {
-      return { success: false, error: ['User not found'], code: 404 };
+      return { success: false, error: ['Error updating user'], code: 404 };
     }
   
     const updatedUserResponse: UpdatedUserResponse = { message: "User updated successfully", userId: data.id }; 
@@ -67,11 +76,11 @@ export async function updateUser(data: UpdateUserRequest): Promise<ServiceResult
   }
 }
 
-export async function getUserById(id: string): Promise<ServiceResult<UserResponse>> {
+export async function getUserByIdAsync(id: string): Promise<ServiceResult<UserResponse>> {
  
   const validationResult = await validate(getUserByIdSchema, {id});  
   if (!validationResult.success) {
-    return validationResult;
+    return { success: false, code: 400, error: validationResult.error }
   }
 
   try { 
@@ -91,11 +100,11 @@ export async function getUserById(id: string): Promise<ServiceResult<UserRespons
   }
 };
  
-export async function getUserByUsername(username: string): Promise<ServiceResult<UserResponse>> {
+export async function getUserByUsernameAsync(username: string): Promise<ServiceResult<UserResponse>> {
  
   const validationResult = await validate(getUserByUsernameSchema, { username });
   if (!validationResult.success) {
-    return validationResult;
+    return { success: false, code: 400, error: validationResult.error }
   } 
 
   try { 
@@ -113,18 +122,21 @@ export async function getUserByUsername(username: string): Promise<ServiceResult
     return handleServiceError(err); 
   }
 }; 
-
  
-export async function deleteUser(id: string): Promise<ServiceResult<DeletedUserResponse>> {
-
+export async function deleteUserAsync(id: string): Promise<ServiceResult<DeletedUserResponse>> {
+  
   const validationResult = await validate(deleteUserSchema, {id});   
   if (!validationResult.success) {
-    return validationResult;
-  }   
+    return { success: false, code: 400, error: validationResult.error }
+  };   
 
   try {  
 
     const { id: validId } = validationResult.data;
+ 
+    if (!(await userExistsAsync(validId))) {
+      return {success: false, code: 404, error: ['User not found']};
+    }   
 
     await UserModel.findByIdAndDelete(validId);
     return { success: true, data: { userId: validId, message: 'User deleted successfully', }, }; 
@@ -134,11 +146,21 @@ export async function deleteUser(id: string): Promise<ServiceResult<DeletedUserR
   } 
 }  
 
-export async function usernameExists(id: string, username: string): Promise<boolean> {
+export async function usernameExistsNotOnThisUserAsync(id: string, username: string): Promise<boolean> {
   const user = await UserModel.findOne({ _id: { $ne: id }, username }).exec();
   return !!user;
 }
 
-export async function getUserByRefreshToken(refreshToken: string): Promise<IUser | null> {
+export async function userExistsAsync(id: string): Promise<boolean> {
+  const user = await UserModel.findOne({ _id: id }).exec();
+  return !!user;
+}
+
+export async function getUserByRefreshTokenAsync(refreshToken: string): Promise<IUser | null> {
   return await UserModel.findOne({ tokens: refreshToken }).exec();
+}
+
+export async function usernameExistsAsync(username: string): Promise<boolean> {
+  const exists = await UserModel.exists({ 'username': username }).exec();
+  return !!exists;
 }
